+++ optd/core/src/ir/operator/logical/order_by.rs
@@ -107,27 +107,27 @@ impl Explain for OrderByBorrowed<'_> {
 #[cfg(test)]
 mod tests {
     use crate::ir::{
-        ColumnSet, IRContext, ScalarValue,
+        ColumnSet, ScalarValue,
         convert::{IntoOperator, IntoScalar},
         scalar::*,
+        table_ref::TableRef,
+        test_utils::{test_col, test_ctx_with_tables},
     };
 
     use super::*;
 
     #[test]
-    fn try_extract_tuple_ordering_success() {
-        let ctx = IRContext::with_empty_magic();
+    fn try_extract_tuple_ordering_success() -> crate::error::Result<()> {
+        let ctx = test_ctx_with_tables(&[("t1", 3)])?;
+        let input = ctx.logical_get(TableRef::bare("t1"), None)?.build();
+        let c0 = test_col(&ctx, "t1", "c0")?;
+        let c1 = test_col(&ctx, "t1", "c1")?;
+        let c2 = test_col(&ctx, "t1", "c2")?;
         let ordered_exprs = vec![
-            (
-                ColumnRef::new(Column(1, 0)).into_scalar(),
-                TupleOrderingDirection::Asc,
-            ),
-            (
-                ColumnRef::new(Column(1, 2)).into_scalar(),
-                TupleOrderingDirection::Asc,
-            ),
+            (ColumnRef::new(c0).into_scalar(), TupleOrderingDirection::Asc),
+            (ColumnRef::new(c2).into_scalar(), TupleOrderingDirection::Asc),
         ];
-        let order_by = OrderBy::new(ctx.mock_scan(1, 3, 100.), ordered_exprs).into_operator();
+        let order_by = OrderBy::new(input, ordered_exprs).into_operator();
 
         let res = order_by
             .borrow::<OrderBy>()
@@ -137,23 +137,27 @@ mod tests {
             .cloned()
             .collect::<ColumnSet>();
 
-        assert!(res.contains(&Column(1, 0)));
-        assert!(res.contains(&Column(1, 2)));
-        assert!(!res.contains(&Column(1, 1)));
+        assert!(res.contains(&c0));
+        assert!(res.contains(&c2));
+        assert!(!res.contains(&c1));
+        Ok(())
     }
 
     #[test]
-    fn try_extract_tuple_ordering_error() {
-        let ctx = IRContext::with_empty_magic();
+    fn try_extract_tuple_ordering_error() -> crate::error::Result<()> {
+        let ctx = test_ctx_with_tables(&[("t1", 3)])?;
+        let input = ctx.logical_get(TableRef::bare("t1"), None)?.build();
+        let c0 = test_col(&ctx, "t1", "c0")?;
+        let c2 = test_col(&ctx, "t1", "c2")?;
         let ordered_exprs = vec![
             (
-                ColumnRef::new(Column(1, 0)).into_scalar(),
+                ColumnRef::new(c0).into_scalar(),
                 TupleOrderingDirection::Asc,
             ),
             (
                 BinaryOp::new(
                     BinaryOpKind::Plus,
-                    ColumnRef::new(Column(1, 2)).into_scalar(),
+                    ColumnRef::new(c2).into_scalar(),
                     Literal::new(ScalarValue::Int32(Some(1))).into_scalar(),
                 )
                 .into_scalar(),
@@ -161,7 +165,7 @@ mod tests {
             ),
         ];
 
-        let order_by = OrderBy::new(ctx.mock_scan(1, 3, 100.), ordered_exprs).into_operator();
+        let order_by = OrderBy::new(input, ordered_exprs).into_operator();
 
         let res = order_by
             .borrow::<OrderBy>()
@@ -169,6 +173,7 @@ mod tests {
             .unwrap_err();
 
         assert_eq!(res.len(), 1);
-        assert!(res[0].try_borrow::<BinaryOp>().is_ok())
+        assert!(res[0].try_borrow::<BinaryOp>().is_ok());
+        Ok(())
     }
 }
