+++ optd/core/src/magic/cm.rs
@@ -18,7 +18,7 @@ impl MagicCostModel {
         Cost::UNIT * card * 2f64
     }
 
-    fn nl_join_cost(
+    fn nested_loop_cost(
         outer_card: crate::ir::properties::Cardinality,
         inner_card: crate::ir::properties::Cardinality,
     ) -> Cost {
@@ -61,16 +61,20 @@ impl CostModel for MagicCostModel {
             OperatorKind::Join(meta) => {
                 let join = Join::borrow_raw_parts(meta, &op.common);
                 match join.implementation() {
-                    Some(implementation) if implementation.is_hash() => {
-                        let build_card = join.build_side().unwrap().cardinality(ctx);
-                        let probe_card = join.probe_side().unwrap().cardinality(ctx);
+                    Some(JoinImplementation::Hash { build_side, .. }) => {
+                        let (build_side, probe_side) = match build_side {
+                            JoinSide::Outer => (join.outer(), join.inner()),
+                            JoinSide::Inner => (join.inner(), join.outer()),
+                        };
+                        let build_card = build_side.cardinality(ctx);
+                        let probe_card = probe_side.cardinality(ctx);
                         let cost = (build_card.as_f64() * 2. + probe_card.as_f64()) * Cost::UNIT;
                         Ok(cost)
                     }
                     _ => {
                         let outer_card = join.outer().cardinality(ctx);
                         let inner_card = join.inner().cardinality(ctx);
-                        Ok(Self::nl_join_cost(outer_card, inner_card))
+                        Ok(Self::nested_loop_cost(outer_card, inner_card))
                     }
                 }
             }
@@ -78,7 +82,7 @@ impl CostModel for MagicCostModel {
                 let join = DependentJoin::borrow_raw_parts(meta, &op.common);
                 let outer_card = join.outer().cardinality(ctx);
                 let inner_card = join.inner().cardinality(ctx);
-                Ok(Self::nl_join_cost(outer_card, inner_card))
+                Ok(Self::nested_loop_cost(outer_card, inner_card))
             }
             OperatorKind::Select(meta) => {
                 let filter = Select::borrow_raw_parts(meta, &op.common);
@@ -126,27 +130,27 @@ impl CostModel for MagicCostModel {
 mod tests {
 
     use crate::ir::{
-        builder_v2::{boolean, column_ref, int64},
+        builder::{boolean, column_ref, int64},
         convert::{IntoOperator, IntoScalar},
         cost::Cost,
         operator::{join::JoinType, *},
         properties::{TupleOrdering, TupleOrderingDirection},
-        scalar::{List, Literal},
+        scalar::Literal,
+        table_ref::TableRef,
+        test_utils::{test_col, test_ctx_with_tables},
         *,
     };
 
     #[test]
-    fn sort_at_lower_cardinality_is_cheaper() {
-        let ctx = IRContext::with_empty_magic();
-
-        let m1 = ctx.mock_scan(1, 2, 100.);
-
-        let m2 = ctx.mock_scan(1, 2, 100.);
+    fn sort_at_lower_cardinality_is_cheaper() -> crate::error::Result<()> {
+        let ctx = test_ctx_with_tables(&[("t1", 2), ("t2", 2)])?;
+        let m1 = ctx.logical_get(TableRef::bare("t1"), None)?.build();
+        let m2 = ctx.logical_get(TableRef::bare("t2"), None)?.build();
+        let t1_c0 = test_col(&ctx, "t1", "c0")?;
 
         let join_cond = Literal::new(ScalarValue::Boolean(Some(true))).into_scalar();
 
-        let tuple_ordering =
-            TupleOrdering::from_iter([(Column(1, 0), TupleOrderingDirection::Asc)]);
+        let tuple_ordering = TupleOrdering::from_iter([(t1_c0, TupleOrderingDirection::Asc)]);
 
         // plan with sort on top.
         let op1 = EnforcerSort::new(
@@ -176,45 +180,37 @@ mod tests {
         let op2_cost = ctx.cm.compute_total_cost(&op2, &ctx).unwrap();
         println!("cost1: {op1_cost:?}, cost2: {op2_cost:?}");
         assert!(op1_cost > op2_cost);
+        Ok(())
     }
 
     #[test]
-    fn logical_cost_matches_physical_counterparts() {
-        let ctx = IRContext::with_empty_magic();
-
-        let schema = catalog::Schema::new(vec![
-            catalog::Field::new("t1.c0", DataType::Int32, false),
-            catalog::Field::new("t1.c1", DataType::Int32, false),
-        ]);
-        let logical_get = ctx.logical_get(1, &schema, None);
-        let physical_scan = ctx.table_scan(1, &schema, None);
+    fn logical_cost_matches_physical_counterparts() -> crate::error::Result<()> {
+        let ctx = test_ctx_with_tables(&[("t1", 2), ("t2", 2)])?;
+        let logical_get = ctx.logical_get(TableRef::bare("t1"), None)?.build();
+        let t1_c0 = test_col(&ctx, "t1", "c0")?;
+        let t1_c1 = test_col(&ctx, "t1", "c1")?;
+        let physical_scan = ctx.table_scan(TableRef::bare("t1"), None)?.build();
         assert_eq!(
-            ctx.cm.compute_operator_cost(&logical_get, &ctx).unwrap(),
-            ctx.cm.compute_operator_cost(&physical_scan, &ctx).unwrap()
+            ctx.cm.compute_operator_cost(&logical_get, &ctx)?,
+            ctx.cm.compute_operator_cost(&physical_scan, &ctx)?
         );
 
-        let outer = ctx.mock_scan(1, 2, 100.);
-        let inner = ctx.mock_scan(2, 2, 50.);
+        let outer = ctx.logical_get(TableRef::bare("t1"), None)?.build();
+        let inner = ctx.logical_get(TableRef::bare("t2"), None)?.build();
         let join_cond = Literal::new(ScalarValue::Boolean(Some(true))).into_scalar();
-        let logical_join = Join::new(
-            JoinType::Inner,
-            outer.clone(),
-            inner.clone(),
-            join_cond.clone(),
-            None,
-        )
-        .into_operator();
-        let physical_join = Join::new(
-            JoinType::Inner,
-            outer.clone(),
-            inner.clone(),
-            join_cond.clone(),
-            Some(JoinImplementation::nested_loop()),
-        )
-        .into_operator();
+        let logical_join = outer
+            .clone()
+            .with_ctx(&ctx)
+            .logical_join(inner.clone(), join_cond.clone(), JoinType::Inner)
+            .build();
+        let physical_join = outer
+            .clone()
+            .with_ctx(&ctx)
+            .nested_loop(inner.clone(), join_cond.clone(), JoinType::Inner)
+            .build();
         assert_eq!(
-            ctx.cm.compute_operator_cost(&logical_join, &ctx).unwrap(),
-            ctx.cm.compute_operator_cost(&physical_join, &ctx).unwrap()
+            ctx.cm.compute_operator_cost(&logical_join, &ctx)?,
+            ctx.cm.compute_operator_cost(&physical_join, &ctx)?
         );
 
         let logical_dep_join = DependentJoin::new(
@@ -225,75 +221,74 @@ mod tests {
         )
         .into_operator();
         assert_eq!(
-            ctx.cm
-                .compute_operator_cost(&logical_dep_join, &ctx)
-                .unwrap(),
-            ctx.cm.compute_operator_cost(&physical_join, &ctx).unwrap()
+            ctx.cm.compute_operator_cost(&logical_dep_join, &ctx)?,
+            ctx.cm.compute_operator_cost(&physical_join, &ctx)?
         );
 
-        let logical_select = Select::new(outer.clone(), join_cond.clone()).into_operator();
-        let physical_filter = Select::new(outer.clone(), join_cond.clone()).into_operator();
+        let logical_select = outer
+            .clone()
+            .with_ctx(&ctx)
+            .select(join_cond.clone())
+            .build();
+        let physical_filter = outer
+            .clone()
+            .with_ctx(&ctx)
+            .select(join_cond.clone())
+            .build();
         assert_eq!(
-            ctx.cm.compute_operator_cost(&logical_select, &ctx).unwrap(),
-            ctx.cm
-                .compute_operator_cost(&physical_filter, &ctx)
-                .unwrap()
+            ctx.cm.compute_operator_cost(&logical_select, &ctx)?,
+            ctx.cm.compute_operator_cost(&physical_filter, &ctx)?
         );
 
-        let projections =
-            List::new(vec![column_ref(Column(1, 0)), column_ref(Column(1, 1))].into())
-                .into_scalar();
-        let logical_project = Project::new(3, outer.clone(), projections.clone()).into_operator();
-        let physical_project = Project::new(3, outer.clone(), projections.clone()).into_operator();
+        let logical_project = ctx
+            .project(outer.clone(), [column_ref(t1_c0), column_ref(t1_c1)])?
+            .build();
+        let physical_project = ctx
+            .project(outer.clone(), [column_ref(t1_c0), column_ref(t1_c1)])?
+            .build();
         assert_eq!(
-            ctx.cm
-                .compute_operator_cost(&logical_project, &ctx)
-                .unwrap(),
-            ctx.cm
-                .compute_operator_cost(&physical_project, &ctx)
-                .unwrap()
+            ctx.cm.compute_operator_cost(&logical_project, &ctx)?,
+            ctx.cm.compute_operator_cost(&physical_project, &ctx)?
         );
 
-        let exprs = List::new(vec![column_ref(Column(1, 1))].into()).into_scalar();
-        let keys = List::new(vec![column_ref(Column(1, 0))].into()).into_scalar();
-        let logical_agg =
-            Aggregate::new(4, outer.clone(), exprs.clone(), keys.clone(), None).into_operator();
-        let physical_agg = Aggregate::new(
-            4,
-            outer.clone(),
-            exprs,
-            keys,
-            Some(AggregateImplementation::Hash),
-        )
-        .into_operator();
+        let logical_agg = outer
+            .clone()
+            .with_ctx(&ctx)
+            .logical_aggregate([column_ref(t1_c1)], [column_ref(t1_c0)])?
+            .build();
+        let physical_agg = outer
+            .clone()
+            .with_ctx(&ctx)
+            .hash_aggregate([column_ref(t1_c1)], [column_ref(t1_c0)])?
+            .build();
         assert_eq!(
-            ctx.cm.compute_operator_cost(&logical_agg, &ctx).unwrap(),
-            ctx.cm.compute_operator_cost(&physical_agg, &ctx).unwrap()
+            ctx.cm.compute_operator_cost(&logical_agg, &ctx)?,
+            ctx.cm.compute_operator_cost(&physical_agg, &ctx)?
         );
 
-        let ordering = TupleOrdering::from_iter([(Column(1, 0), TupleOrderingDirection::Asc)]);
+        let ordering = TupleOrdering::from_iter([(t1_c0, TupleOrderingDirection::Asc)]);
         let logical_order_by = OrderBy::new(
             outer.clone(),
-            vec![(column_ref(Column(1, 0)), TupleOrderingDirection::Asc)],
+            vec![(column_ref(t1_c0), TupleOrderingDirection::Asc)],
         )
         .into_operator();
         let enforcer_sort = EnforcerSort::new(ordering, outer).into_operator();
         assert_eq!(
-            ctx.cm
-                .compute_operator_cost(&logical_order_by, &ctx)
-                .unwrap(),
-            ctx.cm.compute_operator_cost(&enforcer_sort, &ctx).unwrap()
+            ctx.cm.compute_operator_cost(&logical_order_by, &ctx)?,
+            ctx.cm.compute_operator_cost(&enforcer_sort, &ctx)?
         );
+        Ok(())
     }
 
     #[test]
-    fn remaining_logical_operators_have_costs() {
-        let ctx = IRContext::with_empty_magic();
-        let input = ctx.mock_scan(1, 2, 100.);
+    fn remaining_logical_operators_have_costs() -> crate::error::Result<()> {
+        let ctx = test_ctx_with_tables(&[("t1", 2), ("t2", 1)])?;
+        let input = ctx.logical_get(TableRef::bare("t1"), None)?.build();
+        let filter_input = ctx.logical_get(TableRef::bare("t2"), None)?.build();
 
         let limit = Limit::new(input.clone(), int64(10), int64(20)).into_operator();
         let subquery = Subquery::new(input.clone()).into_operator();
-        let remap = Remap::new(2, input).into_operator();
+        let remap = ctx.remap(input, TableRef::bare("t1_remap"))?.build();
 
         assert_eq!(
             ctx.cm.compute_operator_cost(&limit, &ctx).unwrap(),
@@ -310,10 +305,11 @@ mod tests {
         assert!(
             ctx.cm
                 .compute_operator_cost(
-                    &Select::new(ctx.mock_scan(3, 1, 10.), boolean(true)).into_operator(),
+                    &Select::new(filter_input, boolean(true)).into_operator(),
                     &ctx
                 )
                 .is_ok()
         );
+        Ok(())
     }
 }
