+++ optd/core/src/rules/implementations/hash_aggregate.rs
@@ -55,20 +55,25 @@ impl Rule for LogicalAggregateAsPhysicalHashAggregateRule {
 mod tests {
     use super::*;
     use crate::ir::{
-        IRContext,
-        builder_v2::column_ref,
-        convert::{IntoOperator, IntoScalar},
+        builder::column_ref,
         operator::Aggregate,
-        scalar::List,
+        table_ref::TableRef,
+        test_utils::{test_col, test_ctx_with_tables},
     };
 
     #[test]
-    fn lowers_unimplemented_aggregate_to_hash_aggregate() {
-        let ctx = IRContext::with_empty_magic();
-        let input = ctx.mock_scan(1, 2, 10.);
-        let exprs = List::new(vec![column_ref(crate::ir::Column(1, 1))].into()).into_scalar();
-        let keys = List::new(vec![column_ref(crate::ir::Column(1, 0))].into()).into_scalar();
-        let aggregate = Aggregate::new(2, input, exprs.clone(), keys.clone(), None).into_operator();
+    fn lowers_unimplemented_aggregate_to_hash_aggregate() -> crate::error::Result<()> {
+        let ctx = test_ctx_with_tables(&[("t1", 2)])?;
+        let input = ctx.logical_get(TableRef::bare("t1"), None)?.build();
+        let c0 = test_col(&ctx, "t1", "c0")?;
+        let c1 = test_col(&ctx, "t1", "c1")?;
+        let aggregate = input
+            .with_ctx(&ctx)
+            .logical_aggregate([column_ref(c1)], [column_ref(c0)])?
+            .build();
+        let logical_agg = aggregate.try_borrow::<Aggregate>().unwrap();
+        let exprs = logical_agg.exprs().clone();
+        let keys = logical_agg.keys().clone();
 
         let rule = LogicalAggregateAsPhysicalHashAggregateRule::new();
         let transformed = rule.transform(aggregate.as_ref(), &ctx).unwrap();
@@ -81,6 +86,7 @@ mod tests {
         );
         assert_eq!(lowered.exprs(), &exprs);
         assert_eq!(lowered.keys(), &keys);
+        Ok(())
     }
 
     #[test]
