+++ optd/core/src/rules/implementations/table_scan.rs
@@ -53,33 +53,36 @@ mod tests {
     use std::sync::Arc;
 
     use crate::ir::{
-        catalog::DataSourceId,
-        convert::IntoOperator,
         operator::{Get, GetImplementation},
+        table_ref::TableRef,
+        test_utils::test_ctx_with_tables,
     };
 
     use super::*;
 
     #[test]
-    fn logical_get_as_physical_table_scan_behavior() {
-        let logical_get = Get::logical(DataSourceId(1), 1, Arc::new([])).into_operator();
+    fn logical_get_as_physical_table_scan_behavior() -> crate::error::Result<()> {
+        let ctx = test_ctx_with_tables(&[("t1", 1)])?;
+        let logical_get = ctx
+            .logical_get(TableRef::bare("t1"), Some(Arc::new([])))?
+            .build();
+        let logical_get_borrowed = logical_get.try_borrow::<Get>().unwrap();
 
         let rule = LogicalGetAsPhysicalTableScanRule::new();
         assert!(rule.pattern.matches_without_expand(&logical_get));
 
         let after = rule
-            .transform(&logical_get, &crate::ir::IRContext::with_empty_magic())
+            .transform(&logical_get, &ctx)
             .unwrap()
             .pop()
             .unwrap();
         let get = after.try_borrow::<Get>().unwrap();
 
-        assert_eq!(get.data_source_id(), &DataSourceId(1));
-        assert_eq!(get.table_index(), &1);
+        assert_eq!(get.data_source_id(), logical_get_borrowed.data_source_id());
+        assert_eq!(get.table_index(), logical_get_borrowed.table_index());
         assert!(get.projections().is_empty());
         assert_eq!(get.implementation(), &Some(GetImplementation::TableScan));
-
-        let table_scan = Get::table_scan(DataSourceId(1), 1, Arc::new([])).into_operator();
-        assert_eq!(after.kind, table_scan.kind);
+        assert_eq!(get.projections(), logical_get_borrowed.projections());
+        Ok(())
     }
 }
